journal: add monotonic_timestamp()
This function is not only useful for the timestamp itself, but also for
retrieving the boot ID of the current journal entry. IIUC, this is *not*
equivalent to the `_BOOT_ID` entry field, which may not be matched
properly in some cases. For more details, see:
https://github.com/systemd/systemd/commit/dc009662

Implementation-wise, there's nothing surprising here. It's a
straightforward wrapper of the underlying API.  Also add a `from_ffi()`
to convert to the wrapper type from the ffi type and `PartialEq` to help
testing (but which I also need for my use case).