ci: use docker instead of building out own libsystemd
This uses the rustc from the travis environment, but uses libsystemd
from within docker ubuntu:19.04.

This was driven by a need for `sd_bus_slot_set_destroy_callback`, which
wasn't avaliable in the old version of systemd we included as a
submodule. Just plain bumping the submodule to the latest wasn't
possible because systemd now uses meson to build, and current meson
doesn't like ubuntu 14.04.

Docker (or any container env, really) lets us sidestep this issue with
travis.

We need this so the sd-bus code can use the
`sd_bus_slot_set_destroy_callback` to avoid leaking some data when
using callbacks.