features: make them addative (remove elogind, add journal)
For features to work right, they need to be addative. IOW: enabling
features can only _extend_ the API.

elogind does not support the journal apis, so instead provide that as a
feature.

Also, remove the elogind behavir from the build.rs script and document
the use of environment variables to control the build instead.

In the future, we could consider having our build.rs try to link with
elogind when systemd is absent (and features permit it?).