check: install libsystemd-dev so check passes
While we don't currently build (fully) the code here, `cargo check`
still runs our build script, which is unhappy (now that we've fixed it)
if libsystemd's development files (*.pc, mainly) aren't avaliable on the
system.

We could override this (in this case) by setting LIBSYSTEMD_LIBS to
some benign nop (like `libc`), but we eventually want to build, and this
is a good check that the build script works anyhow.