Return custom struct, iterable struct from sd_listen_fds
The documentation and the type of dameon::listen_fds didn't seem to
match.  The type is Fd, but it is used as an integer.  While this isn't
technically a problem, it's a little confusing to read.  This is my
attempt to provide something more useful, and perhaps easier to use
correctly.

This adds two structs: ListenFds and ListenFdsRange.  The first is
returned by listen_fds if assuming there isn't an error.  The second is
returned by the first's iter() method.  This allow for a way to access
multiple file descriptors if they happen to be returned.