Introduce JournalEntry struct and refactor JournalRef to use it
This introduces a new struct, JournalEntry, which just wraps the
JournalRef and then provides an interface to all of the field specific
commands.  I didn't go so far as to deprecate the functionality that was
refactored out without feedback on the general idea of the structure.
This provided a little bit more of a hierarchy that I found easier to
reason about.

This does provide an Iterator implementation, which I made as it's own
struct `EntryIter`, to allow for a different `EntryUniqueIter` and
`EntryUniqueAvailableIter` later on.

I debated about the name of the `next_field` function.  The pattern for
function names in JournalEntry is typically just to remove the
`sd_journal_` and `_data` parts of the C API's function name.  This
would suggest `enumerate` be used.  I worried about that becoming
confusing with `journal.entry().iter().enumerate()`.  Happy to change
that, though, to remain consistent.

Since this is just a refactor, I didn't add any additional tests
currently. I assumed that the coverage on the old functionality (now
using the new functionality under the hood) was sufficient.

This change should be non-breaking, but let me know if I missed
something.